install navigation packages and separate home screen to a component
